chore: Revamp the codebase (#29)
* chore: split files and rearrange codebase

* chore: css files were inverted
