+++ .github/workflows/auto-pr.yml
@@ -10,7 +10,7 @@ jobs:
     steps:
       - uses: actions/checkout@v4
         with:
-          fetch-depth: 0  # Ensures full commit history
+          fetch-depth: 0 # Ensures full commit history
 
       - name: Check for Existing Pull Request
         id: check_pr
