+++ src/App.tsx
@@ -1,16 +1,20 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import GameModeSelector from "./GameModeSelector";
-import Options from './Options';
-import Game from "./Game";
-import { useEffect } from 'react';
-import { HiraganaChar, hiraganaData } from './HiraganaChar';
+import GameModeSelector from "./components/game/GameModeSelector";
+import GameOptions from "./components/game-options/GameOptions";
+import Game from "./components/game/Game";
+import { useEffect } from "react";
+import { HiraganaChar, hiraganaData } from "./services/hiragana-characters";
+
+import "./assets/shared.css";
 
 export default function App() {
   // Initialize localStorage with base hiragana if not exists
   useEffect(() => {
-    if (!localStorage.getItem('selectedHiragana')) {
-      const baseCharacters = hiraganaData.hiragana.map((c: HiraganaChar) => c.character);
-      localStorage.setItem('selectedHiragana', JSON.stringify(baseCharacters));
+    if (!localStorage.getItem("selectedHiragana")) {
+      const baseCharacters = hiraganaData.hiragana.map(
+        (c: HiraganaChar) => c.character,
+      );
+      localStorage.setItem("selectedHiragana", JSON.stringify(baseCharacters));
     }
   }, []);
 
@@ -21,7 +25,7 @@ export default function App() {
         <Route path="/hiragana-quiz" element={<Game mode="hiragana" />} />
         <Route path="/reverse-mode" element={<Game mode="romaji" />} />
         <Route path="/mixed-quiz" element={<Game mode="mixed" />} />
-        <Route path="/options" element={<Options />} />
+        <Route path="/options" element={<GameOptions />} />
       </Routes>
     </Router>
   );
