+++ src/assets/shared.css
@@ -1,0 +1,34 @@
+/* shared.css */
+
+.flipping {
+  transform: rotateY(180deg);
+}
+
+.sparkles {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  pointer-events: none;
+}
+
+.sparkle {
+  position: absolute;
+  width: 4px;
+  height: 4px;
+  background-color: #ffd700;
+  border-radius: 50%;
+  animation: sparkle-animation 0.5s ease-out forwards;
+}
+
+@keyframes sparkle-animation {
+  0% {
+    opacity: 1;
+    transform: scale(1);
+  }
+  100% {
+    opacity: 0;
+    transform: scale(0.5);
+  }
+}
