+++ src/components/game/Game.css
@@ -1,0 +1,65 @@
+/* Game.css */
+.game-container {
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  padding: 20px;
+  align-items: center;
+  overflow-x: hidden; /* Évite l'overflow horizontal */
+  box-sizing: border-box; /* Assure que le padding est inclus dans la largeur totale */
+}
+
+.game-header {
+  display: flex;
+  align-items: center;
+  gap: 20px;
+  margin-bottom: 20px;
+  width: 100%;
+  justify-content: center;
+  box-sizing: border-box; /* Assure que le padding est inclus dans la largeur totale */
+}
+
+.game-back-button {
+  width: 48px;
+  height: 48px;
+  border-radius: 50%;
+  border: none;
+  background: #f0f0f0;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.game-back-button:hover {
+  transform: scale(1.1);
+}
+
+.game-quiz-container {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  box-sizing: border-box; /* Assure que le padding est inclus dans la largeur totale */
+}
+
+/* Media query for mobile devices */
+@media (max-width: 768px) {
+  .game-header {
+    flex-direction: column;
+    align-items: flex-start;
+  }
+
+  .game-back-button {
+    position: absolute;
+    top: 20px;
+    left: 20px;
+  }
+
+  .game-quiz-container {
+    padding: 0 20px;
+  }
+}
