+++ src/components/game/GameModeSelector.css
@@ -1,0 +1,88 @@
+/* GameModeSelector.css */
+.mode-selector-container {
+  position: relative;
+  padding: 20px;
+  text-align: center;
+  overflow-x: hidden; /* Prevent horizontal overflow */
+  box-sizing: border-box; /* Ensure padding is included in the total width */
+}
+
+.options-button {
+  position: absolute;
+  top: 20px;
+  right: 20px;
+  width: 48px;
+  height: 48px;
+  border-radius: 50%;
+  border: none;
+  background: #f0f0f0;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  transition:
+    top 0.3s,
+    bottom 0.3s; /* Smooth transition for position change */
+}
+
+.options-button:hover {
+  transform: scale(1.1);
+}
+
+.mode-grid {
+  display: grid;
+  gap: 20px;
+  grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
+  width: 100%; /* Ensure the container takes the full available width */
+  box-sizing: border-box; /* Ensure padding is included in the total width */
+}
+
+.mode-card {
+  padding: 20px;
+  border-radius: 8px;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+  cursor: pointer;
+  transition: transform 0.2s;
+  box-sizing: border-box; /* Ensure padding is included in the total width */
+}
+
+.mode-card:hover {
+  transform: scale(1.05);
+}
+
+.mode-title {
+  font-size: 1.5rem;
+  margin-bottom: 10px;
+}
+
+.mode-description {
+  font-size: 1rem;
+  color: #555;
+}
+
+/* Media query for mobile devices */
+@media (max-width: 768px) {
+  .mode-grid {
+    grid-template-columns: 1fr; /* Ensure elements are in a single column */
+  }
+
+  .mode-card {
+    padding: 30px;
+  }
+
+  .mode-title {
+    font-size: 1.75rem;
+  }
+
+  .mode-description {
+    font-size: 1.25rem;
+  }
+
+  .options-button {
+    position: fixed; /* Fix the button to the viewport */
+    top: auto; /* Reset top position */
+    bottom: 20px; /* Position at the bottom */
+    right: 20px;
+  }
+}
