+++ src/components/quiz-card/QuizCard.css
@@ -1,0 +1,160 @@
+/* QuizCard.css */
+/* QuizEngine.css */
+.card-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.hiragana-card {
+    width: 100%;
+    min-width: 300px;
+    max-width: 500px;
+    margin: 0 auto;
+    perspective: 1000px;
+}
+
+.card-content {
+    width: 100%;
+    height: 100%;
+    background: rgba(255, 255, 255, 0.95);
+    border-radius: 20px;
+    box-shadow: 0 8px 30px rgba(0, 0, 0, 0.2);
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    padding: 30px;
+    position: relative;
+    backdrop-filter: blur(5px);
+    box-sizing: border-box;
+}
+
+.streak-counter {
+    position: absolute;
+    top: 20px;
+    right: 20px;
+    color: #333;
+    padding: 8px 15px;
+    border-radius: 15px;
+    font-size: 14px;
+    font-weight: bold;
+    box-shadow: 0 0 2px rgba(0, 0, 0, 0.5);
+}
+
+.hiragana {
+    height: 200px;
+    font-size: 120px;
+    color: #333;
+    text-align: center;
+    margin-top: 40px;
+    font-family: "Noto Sans JP", sans-serif;
+    user-select: none;
+    translate: none; /* Prevent auto-translation */
+}
+
+.options-container {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    gap: 10px;
+    margin-top: 0;
+    margin-bottom: 20px;
+}
+
+.option-button {
+    padding: 12px;
+    border: none;
+    border-radius: 10px;
+    background: #f0f0f0;
+    color: #333;
+    font-size: 16px;
+    cursor: pointer;
+    transition: all 0.2s ease;
+    font-family: "Noto Sans JP", sans-serif;
+}
+
+.option-button:hover:not(:disabled) {
+    background: #e0e0e0;
+    transform: translateY(-2px);
+}
+
+.option-button:disabled {
+    cursor: not-allowed;
+    opacity: 0.7;
+}
+
+.option-button.correct {
+    background: #4caf50;
+    color: white;
+}
+
+.option-button.incorrect {
+    background: #f44336;
+    color: white;
+}
+
+@keyframes sparkle {
+    0% {
+        transform: scale(0) rotate(0deg);
+        opacity: 0;
+    }
+    50% {
+        transform: scale(1.2) rotate(180deg);
+        opacity: 1;
+    }
+    100% {
+        transform: scale(0) rotate(360deg);
+        opacity: 0;
+    }
+}
+
+.sparkles {
+    position: absolute;
+    pointer-events: none;
+    width: 100%;
+    height: 100%;
+    left: 0;
+    top: 0;
+}
+
+.sparkle {
+    position: absolute;
+    width: 10px;
+    height: 10px;
+    background: gold;
+    border-radius: 50%;
+    animation: sparkle 0.8s ease-in-out forwards;
+}
+
+@keyframes wrongShake {
+    0%,
+    100% {
+        transform: translateX(0);
+    }
+    25% {
+        transform: translateX(-10px);
+    }
+    75% {
+        transform: translateX(10px);
+    }
+}
+
+.wrong-answer {
+    animation: wrongShake 0.4s ease-in-out;
+    background: rgba(244, 67, 54, 0.1);
+}
+
+.flipping {
+    animation: flipCard 0.3s ease-in-out;
+}
+
+@keyframes flipCard {
+    0% {
+        transform: rotateY(0);
+    }
+    50% {
+        transform: rotateY(90deg);
+    }
+    100% {
+        transform: rotateY(0);
+    }
+}
