+++ tsconfig.json
@@ -1,7 +1,8 @@
 {
-  "include": ["./src/**/*"],
+  "include": ["./src/**/*", "src/svg.d.ts"],
   "compilerOptions": {
     "strict": true,
+    "resolveJsonModule": true,
     "esModuleInterop": true,
     "lib": ["dom", "es2016"],
     "jsx": "react-jsx"
